refactor(mihomo): extract connection_manager and refactor websocket handling
