+++ .cargo/config.toml
@@ -1,2 +1,3 @@
 [env]
-# TS_RS_EXPORT_DIR = { value = "crates/tauri-plugin-mihomo/guest-js/bindings/", relative = true }
+TS_RS_EXPORT_DIR = { value = "crates/tauri-plugin-mihomo/guest-js/bindings/", relative = true }
+TS_RS_LARGE_INT = "number"
